// src/models/interfaces.ts
interface Task {
  id: number;
  text: string;
  completed: boolean;
  failed: boolean;
}

interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}

interface ProductsResponse {
  products: Product[];
  total: number;
  skip: number;
  limit: number;
}

export type { Task, Product, ProductsResponse }